+++ resources/js/Components/CMS/Webpage/SeeAlso1/SeeAlso1Workshop.vue
@@ -64,7 +64,7 @@ const compSwiperOptions = computed(() => {
 </script>
 
 <template>
-  <div id="see-also-carousel" class="w-full pb-6" :style="{
+  <div id="see-also-1-workshop" class="w-full pb-6" :style="{
     ...getStyles(layout?.app?.webpage_layout?.container?.properties, screenType),
     ...getStyles(modelValue.container?.properties, screenType),
     width: 'auto'
